Major Update: Added Projects
Added two projects I worked on together with thorough descriptions.
Implemented simple gallery viewer for the projects and added images for the projects.

Made all images of website round, and added round padding for it to zoom properly when hovered over.

Made my name at the top clickable to return to index.

Added different font for italic parts in headlines.